+++ packages/ssr-runtime/vitest.config.ts
@@ -4,6 +4,11 @@ export default defineConfig({
   test: {
     globals: true,
     environment: "node",
+    typecheck: {
+      enabled: true,
+      include: ["tests/**/*.test-d.ts"],
+      tsconfig: "./tsconfig.test.json",
+    },
     coverage: {
       provider: "v8",
       reporter: ["text", "json", "html"],